+++ mail/mail-common/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Available actions for conversations in bottom bar.
 - `GeneralActions::ViewMessageInDarkMode` general action.
 - `LabelAs` action for conversations.
 - Available actions for messages in bottom bar.